Make Dektec DTAPI inclusion optional
Add an "--enable-dtapi" argument to allow the user to exclude
the DTAPI from their build.  For now it defaults to true for
backward compatibility, but the build scripts should probably be
tweaked and/or we should have an "auto" option that sees if
it's in the environment and working properly.