Handle 7721DE embedder payloads where the GPI pins are active
We previously reverse engineered the serial protocol encapsulation
in the Evertz 7721DE embedder, and during our testing we found that
there was a common four byte header:  "18 00 c7 00".  We figured out
what the first three bytes are (serial parameters like baud rate,
stop bits, etc), but the last byte was always 0x00, so when we
implemented the parser we bail out if it's ever anything other than
that value.

It turns out that byte actually contains the value of the various
GPI pins on the embedder, and it seems like some of the NBA venues
are using those pins (for what, I have no idea).  As a result, when
any of the GPI pins are active that last byte is something other
than 0x00 and we were discarding the packets.

Relax the header check to allow the fourth byte to be any value.